# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hakimi Healthcare API
  version: 1.0.0
  license:
    name: UNLICENSED
    identifier: UNLICENSED
  description: |
    REST API for Hakimi / ሀኪሜ, a healthcare appointment platform for Ethiopia.

    Notes:
    - Workforce API operations require an OIDC bearer access token with workforce MFA.
    - Authorization is default deny and derives roles and facility scope from server-side PostgreSQL state.
    - Patient authentication and self-service remain unavailable.
    - `DELETE` on facilities, practitioners, and assignments performs a soft delete.
    - Practitioner facility primary reassignment is handled atomically inside one transaction.
servers:
  - url: http://127.0.0.1:3001
security:
  - workforceBearer: []
...
paths:
...
  /api/v1/facilities:
    post:
      tags: [Facilities]
      summary: Create a healthcare facility
      operationId: createHealthcareFacility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHealthcareFacilityInput'
            examples:
              clinic:
                summary: Fictional clinic
                value:
                  code: ADD-CLINIC-001
                  name: Addis Sunrise Clinic
                  facilityType: clinic
                  licenseNumber: LIC-ADD-001
                  phone: '+251911111111'
                  email: contact@addissunrise.example
                  region: Addis Ababa
                  city: Addis Ababa
                  addressLine: Bole Road
                  isActive: true
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '201':
          description: Created facility
          headers:
            Location:
              description: URL of the created facility
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthcareFacility'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/FacilityConflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Facilities]
      summary: List healthcare facilities
      operationId: listHealthcareFacilities
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: facilityType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/FacilityType'
        - name: region
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: city
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
            minLength: 2
            maxLength: 100
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Paged list of facilities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthcareFacilityListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/practitioners:
    post:
      tags: [Practitioners]
      summary: Create a practitioner
      operationId: createPractitioner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePractitionerInput'
            examples:
              physician:
                summary: Fictional practitioner
                value:
                  code: PRAC-001
                  firstName: Mekdes
                  middleName: A.
                  lastName: Tadesse
                  profession: general practitioner
                  licenseNumber: MED-001
                  phone: '+251911111111'
                  email: mekdes@example.org
                  bio: Integration test practitioner
                  isActive: true
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '201':
          description: Created practitioner
          headers:
            Location:
              description: URL of the created practitioner
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Practitioner'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/PractitionerConflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Practitioners]
      summary: List practitioners
      operationId: listPractitioners
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: profession
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
        - name: facilityId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          required: false
          description: Case-insensitive search across code, names, profession, and license number.
          schema:
            type: string
            minLength: 2
            maxLength: 100
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Paged practitioner list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PractitionerListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/patients:
    post:
      tags: [Patients]
      summary: Create a patient with an initial facility registration
      description: |
        Creates the patient demographics row and the first facility registration in a single transaction.
        Medical record numbers are unique per facility, not globally.
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientInput'
            examples:
              registration:
                summary: Create a registered patient
                value:
                  facilityId: 33333333-3333-4333-8333-333333333333
                  medicalRecordNumber: MRN-0001
                  firstName: Abebe
                  middleName: A.
                  lastName: Kebede
                  dateOfBirth: '1994-06-12'
                  administrativeSex: male
                  phone: '+251911111111'
                  email: abebe@example.org
                  addressLine: Bole Road
                  city: Addis Ababa
                  region: Addis Ababa
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '201':
          description: Created patient
          headers:
            Location:
              description: URL of the created patient
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/FacilityNotFound'
        '409':
          $ref: '#/components/responses/PatientCreateConflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Patients]
      summary: List patients
      operationId: listPatients
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          required: false
          description: Case-insensitive search across names, contact fields, and MRN.
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - name: facilityId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: medicalRecordNumber
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 50
        - name: administrativeSex
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AdministrativeSex'
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Paged patient list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/appointments:
    post:
      tags: [Appointments]
      summary: Create an appointment
      description: |
        Creates an appointment in a single transaction after confirming the facility, practitioner, and patient
        are active and linked correctly. Only non-cancelled appointments block a practitioner time slot.
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentInput'
            examples:
              schedule:
                summary: Schedule a new appointment
                value:
                  patientId: 11111111-1111-4111-8111-111111111111
                  practitionerId: 22222222-2222-4222-8222-222222222222
                  facilityId: 33333333-3333-4333-8333-333333333333
                  scheduledStart: '2026-08-07T09:00:00+03:00'
                  scheduledEnd: '2026-08-07T09:30:00+03:00'
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '201':
          description: Created appointment
          headers:
            Location:
              description: URL of the created appointment
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/AppointmentCreateLookupError'
        '409':
          $ref: '#/components/responses/AppointmentCreateConflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Appointments]
      summary: List appointments
      description: |
        Returns appointments in deterministic scheduled-start order. Date range filters use overlapping windows so
        appointments partially inside the requested window are included.
      operationId: listAppointments
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: facilityId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: practitionerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: patientId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AppointmentStatus'
        - name: from
          in: query
          required: false
          description: Inclusive lower bound for overlapping appointment windows.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Exclusive upper bound for overlapping appointment windows.
          schema:
            type: string
            format: date-time
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Paged appointment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    workforceBearer:
      type: oauth2
      description: |
        Workforce-only OIDC Authorization Code with PKCE. The resource server
        requires a short-lived signed JWT access token, approved workforce MFA,
        and an active server-provisioned actor, role, scope, and local session.
        Patient accounts and role-administration APIs are not available.
      flows:
        authorizationCode:
          authorizationUrl: https://identity.example.com/oauth2/authorize
          tokenUrl: https://identity.example.com/oauth2/token
          scopes: {}
  schemas:
    FacilityType:
      type: string
      enum:
        - hospital
        - clinic
        - health_center
        - diagnostic_center
        - pharmacy
    HealthcareFacility:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - name
        - facilityType
        - licenseNumber
        - phone
        - email
        - region
        - city
        - addressLine
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        facilityType:
          $ref: '#/components/schemas/FacilityType'
        licenseNumber:
          type:
            - string
            - 'null'
        phone:
          type:
            - string
            - 'null'
        email:
          type:
            - string
            - 'null'
        region:
          type: string
        city:
          type: string
        addressLine:
          type:
            - string
            - 'null'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateHealthcareFacilityInput:
      type: object
      additionalProperties: false
      required:
        - code
        - name
        - facilityType
        - region
        - city
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 50
        name:
          type: string
          minLength: 2
          maxLength: 200
        facilityType:
          $ref: '#/components/schemas/FacilityType'
        licenseNumber:
          type:
            - string
            - 'null'
          maxLength: 100
        phone:
          type:
            - string
            - 'null'
          maxLength: 30
        email:
          type:
            - string
            - 'null'
          format: email
          maxLength: 254
        region:
          type: string
          minLength: 2
          maxLength: 100
        city:
          type: string
          minLength: 2
          maxLength: 100
        addressLine:
          type:
            - string
            - 'null'
          maxLength: 300
        isActive:
          type: boolean
          default: true
...
    HealthcareFacilityPagination:
      type: object
      additionalProperties: false
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    HealthcareFacilityListResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/HealthcareFacility'
        pagination:
          $ref: '#/components/schemas/HealthcareFacilityPagination'
    Practitioner:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - firstName
        - middleName
        - lastName
        - profession
        - licenseNumber
        - phone
        - email
        - bio
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        firstName:
          type: string
        middleName:
          type:
            - string
            - 'null'
        lastName:
          type: string
        profession:
          type: string
        licenseNumber:
          type: string
        phone:
          type:
            - string
            - 'null'
        email:
          type:
            - string
            - 'null'
        bio:
          type:
            - string
            - 'null'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePractitionerInput:
      type: object
      additionalProperties: false
      required:
        - code
        - firstName
        - lastName
        - profession
        - licenseNumber
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 50
        firstName:
          type: string
          minLength: 2
          maxLength: 100
        middleName:
          type:
            - string
            - 'null'
          maxLength: 100
        lastName:
          type: string
          minLength: 2
          maxLength: 100
        profession:
          type: string
          minLength: 2
          maxLength: 100
        licenseNumber:
          type: string
          minLength: 2
          maxLength: 100
        phone:
          type:
            - string
            - 'null'
          maxLength: 30
        email:
          type:
            - string
            - 'null'
          format: email
          maxLength: 254
        bio:
          type:
            - string
            - 'null'
          maxLength: 2000
        isActive:
          type: boolean
          default: true
...
    PractitionerListResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Practitioner'
        pagination:
          $ref: '#/components/schemas/PractitionerPagination'
...
    AdministrativeSex:
      type: string
      enum:
        - female
        - male
        - other
        - unknown
...
    Patient:
      type: object
      additionalProperties: false
      required:
        - id
        - firstName
        - middleName
        - lastName
        - dateOfBirth
        - administrativeSex
        - phone
        - email
        - addressLine
        - city
        - region
        - isActive
        - createdAt
        - updatedAt
        - registrations
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        middleName:
          type:
            - string
            - 'null'
        lastName:
          type:
            - string
            - 'null'
        dateOfBirth:
          type:
            - string
            - 'null'
          format: date
        administrativeSex:
          $ref: '#/components/schemas/AdministrativeSex'
        phone:
          type:
            - string
            - 'null'
        email:
          type:
            - string
            - 'null'
        addressLine:
          type:
            - string
            - 'null'
        city:
          type:
            - string
            - 'null'
        region:
          type:
            - string
            - 'null'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        registrations:
          type: array
          items:
            $ref: '#/components/schemas/PatientFacilityRegistration'
    CreatePatientInput:
      type: object
      additionalProperties: false
      required:
        - facilityId
        - medicalRecordNumber
        - firstName
        - administrativeSex
      properties:
        facilityId:
          type: string
          format: uuid
        medicalRecordNumber:
          type: string
          minLength: 1
          maxLength: 50
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        middleName:
          type:
            - string
            - 'null'
          maxLength: 100
        lastName:
          type:
            - string
            - 'null'
          maxLength: 100
        dateOfBirth:
          type:
            - string
            - 'null'
          format: date
        administrativeSex:
          $ref: '#/components/schemas/AdministrativeSex'
        phone:
          type:
            - string
            - 'null'
          maxLength: 30
        email:
          type:
            - string
            - 'null'
          format: email
          maxLength: 254
        addressLine:
          type:
            - string
            - 'null'
          maxLength: 200
        city:
          type:
            - string
            - 'null'
          maxLength: 100
        region:
          type:
            - string
            - 'null'
          maxLength: 100
...
    PatientListResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Patient'
        pagination:
          $ref: '#/components/schemas/PatientPagination'
    AppointmentStatus:
      type: string
      enum:
        - SCHEDULED
        - CONFIRMED
        - COMPLETED
        - CANCELLED
        - NO_SHOW
...
    Appointment:
      type: object
      additionalProperties: false
      required:
        - id
        - patientId
        - practitionerId
        - facilityId
        - scheduledStart
        - scheduledEnd
        - status
        - cancellationReason
        - cancelledAt
        - createdAt
        - updatedAt
        - patient
        - practitioner
        - facility
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        practitionerId:
          type: string
          format: uuid
        facilityId:
          type: string
          format: uuid
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        cancellationReason:
          type:
            - string
            - 'null'
        cancelledAt:
          type:
            - string
            - 'null'
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        patient:
          $ref: '#/components/schemas/AppointmentPatientSummary'
        practitioner:
          $ref: '#/components/schemas/AppointmentPractitionerSummary'
        facility:
          $ref: '#/components/schemas/AppointmentFacilitySummary'
    CreateAppointmentInput:
      type: object
      additionalProperties: false
      required:
        - patientId
        - practitionerId
        - facilityId
        - scheduledStart
        - scheduledEnd
      properties:
        patientId:
          type: string
          format: uuid
        practitionerId:
          type: string
          format: uuid
        facilityId:
          type: string
          format: uuid
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
...
    AppointmentListResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
        pagination:
          $ref: '#/components/schemas/AppointmentPagination'
...
  responses:
    AuthenticationRequired:
      description: Missing, invalid, expired, revoked, or otherwise ineligible workforce authentication
      headers:
        WWW-Authenticate:
          description: Bearer authentication challenge
          schema:
            type: string
            const: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            authenticationRequired:
              value:
                error:
                  code: AUTHENTICATION_REQUIRED
                  message: Authentication required
    Forbidden:
      description: Authenticated workforce actor is not permitted to perform this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            forbidden:
              value:
                error:
                  code: FORBIDDEN
                  message: Forbidden
    ValidationError:
      description: Validation failure or malformed request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            validation:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Request validation failed
                  details:
                    - field: name
                      message: Name is required
    FacilityNotFound:
      description: Facility not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            notFound:
              value:
                error:
                  code: FACILITY_NOT_FOUND
                  message: Facility not found
    FacilityConflict:
      description: Facility code or license number conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            codeConflict:
              value:
                error:
                  code: FACILITY_CODE_CONFLICT
                  message: Facility code already exists
            licenseConflict:
              value:
                error:
                  code: FACILITY_LICENSE_CONFLICT
                  message: License number already exists
...
    PractitionerConflict:
      description: Practitioner code or license number conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            codeConflict:
              value:
                error:
                  code: PRACTITIONER_CODE_CONFLICT
                  message: Practitioner code already exists
            licenseConflict:
              value:
                error:
                  code: PRACTITIONER_LICENSE_CONFLICT
                  message: Practitioner license number already exists
...
    PatientCreateConflict:
      description: Patient creation can fail because the facility is inactive or the MRN already exists at that facility.
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: FACILITY_INACTIVE
                      message:
                        type: string
                        const: Facility is inactive
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: PATIENT_REGISTRATION_CONFLICT
                      message:
                        type: string
                        const: Patient registration already exists
          examples:
            inactiveFacility:
              summary: Facility is inactive
              value:
                error:
                  code: FACILITY_INACTIVE
                  message: Facility is inactive
            duplicateRegistration:
              summary: Duplicate MRN at facility
              value:
                error:
                  code: PATIENT_REGISTRATION_CONFLICT
                  message: Patient registration already exists
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            internal:
              value:
                error:
                  code: INTERNAL_ERROR
                  message: Internal server error